update sql query

+++ src/main/resources/db/migration/V1__Create_all_skills.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS skills (
   id  int(11) NOT NULL AUTO_INCREMENT,
   name varchar(45) NOT NULL,
-  status int(4) NOT NULL,
+  status SMALLINT NOT NULL,
   PRIMARY KEY (id)
 )
