+++ src/main/resources/db/migration/V2__Create_specialty_table.sql
@@ -2,7 +2,7 @@
 CREATE TABLE IF NOT EXISTS specialties (
   id  int(11) NOT NULL AUTO_INCREMENT,
   name varchar(45) NOT NULL,
-  status int(4) NOT NULL,
+  status SMALLINT NOT NULL,
   PRIMARY KEY (id)
 )
